implemented fixtures for money and denom
